Move resetGoals into extension pack
